+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
   "crates/kinematics",
   "crates/nao",
   "crates/nao_camera",
+  "crates/parameters",
   "crates/projection",
   "crates/motionfile",
   "crates/repository",
@@ -100,6 +101,7 @@ petgraph = "0.6.2"
 png = "0.17.6"
 proc-macro-error = "1.0.4"
 proc-macro2 = { version = "1.0.44", features = ["span-locations"] }
+parameters = { path = "crates/parameters" }
 projection = { path = "crates/projection" }
 quote = "1.0.21"
 rand = "0.8.5"
