+++ crates/communication/Cargo.toml
@@ -13,6 +13,7 @@ color-eyre = { workspace = true }
 framework = { workspace = true, optional = true}
 futures-util = { workspace = true }
 log = { workspace = true }
+parameters = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serialize_hierarchy = { workspace = true }
