+++ crates/communication/src/server/parameters/mod.rs
@@ -4,7 +4,6 @@ use crate::messages::Path;
 
 use super::client::Client;
 
-pub mod directory;
 pub mod storage;
 pub mod subscriptions;
