+++ crates/parameters/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "parameters"
+version = "0.1.0"
+edition = "2021"
+license = "GPL-3.0-only"
+homepage = "https://github.com/hulks/hulk"
+
+[dependencies]
+log = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+serialize_hierarchy = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true }
